finisched refactoring
