+++ sputofy_2/lib/main.dart
@@ -8,7 +8,6 @@ import 'package:provider/provider.dart';
 import 'providers/provider.dart';
 
 import 'package:permission_handler/permission_handler.dart';
-import 'package:share_plus/share_plus.dart';
 
 import 'screens/MiniPlayerScreen/mini_player.dart';
 import 'screens/playlistListScreen/playlist_list_screen.dart';
@@ -19,12 +18,6 @@ import 'theme/style.dart';
 import 'components/download_song.dart';
 import 'components/load_song.dart';
 import 'components/playlist_dialog.dart';
-
-import 'models/playlist_model.dart';
-import 'models/song_model.dart';
-
-import 'screens/EditPlaylistScreen/edit_playlist_screen.dart';
-import 'screens/EditSongScreen/edit_song_screen.dart';
 
 void main() {
   runApp(MyApp());
@@ -53,7 +46,6 @@ class MyApp extends StatelessWidget {
 }
 
 _backgroundTaskEntryPoint() {
-  print("entrypoint");
   AudioServiceBackground.run(() => AudioPlayerTask());
 }
 
@@ -68,11 +60,11 @@ class MyHomePage extends StatefulWidget {
 
 class _MyHomePageState extends State<MyHomePage> {
   int tabIndex = 0;
+  //! SearchBar related
   bool isSearching = false;
   Widget appBarTitle = Text('Sputofy');
   @override
   void initState() {
-    print("Initi");
     Permission.storage.request();
 
     _start();
@@ -86,9 +78,6 @@ class _MyHomePageState extends State<MyHomePage> {
       androidEnableQueue: true,
       androidNotificationIcon: 'mipmap/ic_launcher',
       androidStopForegroundOnPause: true,
-
-      // androidNotificationColor: 0x00000000,
-      // androidNotificationColor: 0x0000ff,
     );
   }
 
@@ -103,6 +92,7 @@ class _MyHomePageState extends State<MyHomePage> {
           case 0:
             setState(() {
               tabIndex = 0;
+              //! SearchBar related
               // isSearching = false;
               // appBarTitle = Text('Sputofy');
             });
@@ -110,6 +100,7 @@ class _MyHomePageState extends State<MyHomePage> {
           case 1:
             setState(() {
               tabIndex = 1;
+              //! SearchBar related
               // isSearching = false;
               // appBarTitle = Text('Sputofy');
             });
@@ -122,62 +113,80 @@ class _MyHomePageState extends State<MyHomePage> {
       appBar: AppBar(
         title: appBarTitle,
         actions: <Widget>[
-          if (!isSearching) ...[
-            IconButton(
-              onPressed: () {
-                setState(() {
-                  isSearching = true;
-                  appBarTitle = TextField();
-                });
-              },
-              icon: Icon(Icons.search),
-            ),
-            tabIndex == 0
-                ? PopupMenuButton<String>(
-                    onSelected: (String choice) =>
-                        _handleClick(choice, context),
-                    itemBuilder: (BuildContext context) {
-                      return {'Download Song', 'Load Songs'}
-                          .map((String choice) {
-                        return PopupMenuItem<String>(
-                          value: choice,
-                          child: Text(choice),
-                          // child: Row(
-                          // children: [
-                          //   Text("62.5%"),
-                          //   SizedBox(width: 4.0),
-                          //   Text(
-                          //     choice,
-                          //   ),
-                          // ],
-                          // ),
-                        );
-                      }).toList();
-                    },
-                  )
-                : IconButton(
-                    onPressed: () => showNewPlaylistDialog(context),
-                    icon: Icon(Icons.add),
-                  ),
-          ] else ...[
-            GestureDetector(
-              onTap: () {
-                setState(() {
-                  isSearching = false;
-                  appBarTitle = Text('Sputofy');
-                });
-              },
-              child: Text("cacca"),
-            )
-          ],
+          IconButton(
+            onPressed: () {
+              ScaffoldMessenger.of(context).showSnackBar(
+                SnackBar(
+                  content: Text('Not implemented yet'),
+                ),
+              );
+            },
+            icon: Icon(Icons.search),
+          ),
+          tabIndex == 0
+              ? PopupMenuButton<String>(
+                  onSelected: (String choice) => _handleClick(choice, context),
+                  itemBuilder: (BuildContext context) {
+                    return {'Download Song', 'Load Songs'}.map((String choice) {
+                      return PopupMenuItem<String>(
+                        value: choice,
+                        child: Text(choice),
+                      );
+                    }).toList();
+                  },
+                )
+              : IconButton(
+                  onPressed: () => showNewPlaylistDialog(context),
+                  icon: Icon(Icons.add),
+                ),
+          //! Uncomment for enabling search bar (needs to be implemented)
+          // if (!isSearching) ...[
+          //   IconButton(
+          //     onPressed: () {
+          //       setState(() {
+          //         isSearching = true;
+          //         appBarTitle = TextField();
+          //       });
+          //     },
+          //     icon: Icon(Icons.search),
+          //   ),
+          //   tabIndex == 0
+          //       ? PopupMenuButton<String>(
+          //           onSelected: (String choice) =>
+          //               _handleClick(choice, context),
+          //           itemBuilder: (BuildContext context) {
+          //             return {'Download Song', 'Load Songs'}
+          //                 .map((String choice) {
+          //               return PopupMenuItem<String>(
+          //                 value: choice,
+          //                 child: Text(choice),
+          //               );
+          //             }).toList();
+          //           },
+          //         )
+          //       : IconButton(
+          //           onPressed: () => showNewPlaylistDialog(context),
+          //           icon: Icon(Icons.add),
+          //         ),
+          // ] else ...[
+          //   GestureDetector(
+          //     onTap: () {
+          //       setState(() {
+          //         isSearching = false;
+          //         appBarTitle = Text('Sputofy');
+          //       });
+          //     },
+          //     child: Text("cancel"),
+          //   )
+          // ],
         ],
         bottom: TabBar(
           tabs: [
             Tab(
-              text: "Songs",
+              text: 'Songs',
             ),
             Tab(
-              text: "Playlist",
+              text: 'Playlist',
             )
           ],
         ),
@@ -191,7 +200,7 @@ class _MyHomePageState extends State<MyHomePage> {
             children: [
               SongListScreen(),
               PlaylistListScreen(),
-              //* Uncomment if you want to display something while is loading AudioService Isolate
+              //! Uncomment if you want to display something while is loading AudioService Isolate
               // if (snapshot.connectionState != ConnectionState.active) ...[
               //   // SizedBox(),
               //   Container(
@@ -225,11 +234,9 @@ void _handleClick(String choice, BuildContext context) async {
       try {
         final result = await InternetAddress.lookup('example.com');
         if (result.isNotEmpty && result[0].rawAddress.isNotEmpty) {
-          print('connected');
           isConnected = true;
         }
       } on SocketException catch (_) {
-        print('not connected');
         isConnected = false;
       }
       if (isConnected) {
@@ -263,149 +270,4 @@ void _handleClick(String choice, BuildContext context) async {
       loadSongs(context);
       break;
   }
-}
-
-//TODO move to components
-Widget songMenuButton(Song song, BuildContext context,
-    {Playlist? playlist, IconData? icon, bool? shouldPop}) {
-  return PopupMenuButton<List>(
-    onSelected: (List params) => songMenuHandleClick(params, context,
-        playlist: playlist, shouldPop: shouldPop),
-    icon: icon == null ? Icon(Icons.more_horiz) : Icon(icon),
-    padding: EdgeInsets.zero,
-    itemBuilder: (context) {
-      return {'Delete Song', 'Edit Song', 'Share Song'}.map((String choice) {
-        return PopupMenuItem<List>(
-          value: [choice, song],
-          child: Text(choice),
-        );
-      }).toList();
-    },
-  );
-}
-
-songMenuHandleClick(List params, BuildContext context,
-    {Playlist? playlist, bool? shouldPop}) {
-  //* params = [choice, song]
-  String choice = params[0];
-  Song song = params[1];
-
-  switch (choice) {
-    case 'Delete Song':
-      _deleteSong(song, context, playlist: playlist, shouldPop: shouldPop);
-      break;
-    case 'Edit Song':
-      Navigator.push(
-        context,
-        MaterialPageRoute(
-          builder: (context) => EditSongScreen(song, playlist: playlist),
-        ),
-      );
-      break;
-    case 'Share Song':
-      Share.shareFiles([song.path]);
-      break;
-  }
-}
-
-void _deleteSong(Song song, BuildContext context,
-    {Playlist? playlist, bool? shouldPop}) {
-  if (playlist != null) {
-    if (playlist.id == 0) {
-      AudioService.updateMediaItem(
-          song.copyWith(isFavorite: false).toMediaItem());
-      Provider.of<DBProvider>(context, listen: false)
-          .updateSong(song.copyWith(isFavorite: false));
-    } else {
-      if (AudioService.currentMediaItem?.album == '${playlist.id}') {
-        AudioService.removeQueueItem(song.toMediaItem());
-      }
-    }
-    Provider.of<DBProvider>(context, listen: false)
-        .deletePlaylistSong(playlist.id!, song.id!);
-  } else {
-    if (AudioService.currentMediaItem?.album == '-2') {
-      AudioService.removeQueueItem(song.toMediaItem());
-    }
-    Provider.of<DBProvider>(context, listen: false).deleteSong(song.id!);
-  }
-  if (shouldPop != null && shouldPop) {
-    Navigator.pop(context);
-  }
-}
-
-Widget playlistMenuButton(
-    Playlist playlist, BuildContext context, bool shouldPop) {
-  return PopupMenuButton<List>(
-    onSelected: (List params) =>
-        playlistMenuHandleClick(params, context, shouldPop),
-    icon: Icon(Icons.more_vert),
-    padding: EdgeInsets.zero,
-    itemBuilder: (context) {
-      return {'Delete Playlist', 'Edit Playlist'}.map((String choice) {
-        return PopupMenuItem<List>(
-          value: [choice, playlist],
-          child: Text(choice),
-        );
-      }).toList();
-    },
-  );
-}
-
-playlistMenuHandleClick(List params, BuildContext context, bool shouldPop) {
-  //* params = [choice, playlist]
-  String choice = params[0];
-  Playlist playlist = params[1];
-  switch (choice) {
-    case 'Delete Playlist':
-      if (playlist.id == 0) {
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(
-            content: Text("Favorite playlist can't be deleted"),
-            action: SnackBarAction(
-              label: 'HIDE',
-              onPressed: () =>
-                  ScaffoldMessenger.of(context).hideCurrentSnackBar(),
-            ),
-          ),
-        );
-      } else {
-        Provider.of<DBProvider>(context, listen: false)
-            .deletePlaylist(playlist.id!);
-        if (shouldPop) Navigator.pop(context);
-      }
-      break;
-    case 'Edit Playlist':
-      Navigator.push(
-        context,
-        MaterialPageRoute(
-          builder: (context) => EditPlaylistScreen(playlist),
-        ),
-      );
-      break;
-  }
-}
-
-Future<void> loadQueue(int playlistID, List<Song> songs,
-    {String? songPath, String? playlistTitle}) async {
-  if (songs.isEmpty) return;
-  List<MediaItem> mediaItems = [];
-  for (Song song in songs) {
-    mediaItems.add(
-      song
-          .toMediaItem(
-              playlistTitle: playlistTitle != null ? playlistTitle : null)
-          .copyWith(album: '$playlistID'),
-    );
-  }
-  await AudioService.updateQueue(mediaItems).then(
-    (value) => {
-      if (songPath != null)
-        {
-          AudioService.skipToQueueItem(songPath).then(
-            (value) async => await AudioService.play(),
-          )
-        }
-    },
-  );
 }
