+++ sputofy_2/lib/screens/SongDetailScreen/components/top_bar.dart
@@ -1,11 +1,12 @@
-import 'package:audio_service/audio_service.dart';
 import 'package:flutter/material.dart';
+import 'package:audio_service/audio_service.dart';
+
+import 'package:sputofy_2/components/song_menu_button.dart';
 
 import 'package:sputofy_2/models/playlist_model.dart';
 import 'package:sputofy_2/models/song_model.dart';
-import 'package:sputofy_2/theme/palette.dart';
 
-import '../../../main.dart';
+import 'package:sputofy_2/theme/palette.dart';
 
 Widget topBar(BuildContext context, MediaItem playingItem) {
   return Row(
