+++ sputofy_2/lib/screens/playlistSongsScreen/components/song_tile.dart
@@ -1,11 +1,13 @@
 import 'package:audio_service/audio_service.dart';
 import 'package:flutter/material.dart';
 
+import 'package:sputofy_2/components/load_queue.dart';
+import 'package:sputofy_2/components/song_menu_button.dart';
+
 import 'package:sputofy_2/models/playlist_model.dart';
 import 'package:sputofy_2/models/song_model.dart';
-import 'package:sputofy_2/theme/palette.dart';
 
-import '../../../main.dart';
+import 'package:sputofy_2/theme/palette.dart';
 
 class SongTile extends StatelessWidget {
   final Song song;
